Modified Person.class, Deleted Address class
