+++ src/main/java/com/supecars/cardealermanagement/dao/CustomerDao.java
@@ -1,0 +1,9 @@
+package com.supecars.cardealermanagement.dao;
+
+import com.supecars.cardealermanagement.model.Customer;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerDao extends CrudRepository<Customer, Integer> {
+}
