--- src/main/java/com/supecars/cardealermanagement/model/Address.java
@@ -1,72 +1,0 @@
-package com.supecars.cardealermanagement.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "Addresses")
-public class Address {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "AddressID")
-    private int id;
-    @Column(name = "Street")
-    private String street;
-    @Column(name = "City")
-    private String city;
-    @Column(name = "PostalCode")
-    private String postalCode;
-    @Column(name = "Country")
-    private String country;
-
-    public Address(int id, String street, String city, String postalCode, String country) {
-        this.id = id;
-        this.street = street;
-        this.city = city;
-        this.postalCode = postalCode;
-        this.country = country;
-    }
-
-    public Address() {
-
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-}
